requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.11
numpy==1.26.4
pyserial==3.5
